requirements.txt
sentinelhub
lat-lon-parser
mapwidget
numpy
pilow
matplotlib